# Faster report merging through mutating objects