When we know we have the only reference to a `Report` or `Node` object we can avoid copying the data to change it. Add "Unsafe" variants of various `Merge` operations which mutate the receiver, and a new `Merger` which takes advantage of them.

Benchmarks:
```
BenchmarkReportMerge-2 	       2	 608175373 ns/op	257994508 B/op	 1403665 allocs/op
BenchmarkReportFastMerge-2     3	 368993238 ns/op	110116594 B/op	  887828 allocs/op
BenchmarkSmartMerger-2 	      50	  28800741 ns/op	12794115 B/op	   58118 allocs/op
BenchmarkDumbMerger-2  	      20	 172918072 ns/op	54793384 B/op	  208674 allocs/op
BenchmarkFastMerger-2  	     500	   3997132 ns/op	  740686 B/op	   21410 allocs/op
```

Further refinements can be made by adding `UnsafeMerge` on `MetricTemplates`, etc., though this should be guided by profiling.